refactor: extract follow-up queue from AgentWorkspace
